Latest docker image - using new versioning scheme! (#189)
* Latest PiHole Docker version

Quote from https://github.com/pi-hole/docker-pi-hole/releases/tag/2021.09:

Previously, we have picked a version number of one of the internal components (most recently, FTL) and applied that to the docker container's tag. However, this sometimes causes confusion between users, but also developers (when we are not sure what to tag a docker release if FTL's version number has not changed)

Going forward, we will use a year.month[.revision] to tag docker releases with. To tag a new image, it is as simple as looking at the calendar to decide what the tag name should be. We will always try to leave any major changes until a new month rolls over. e.g if Pi-hole v6 were to be released in 2 weeks time, it would not be in the docker image until 2021.10 (that said, we can also tag pre-release versions for those that are itching to get onto the bleeding edge. This also gives us chance to make sure that none of the core Pi-hole changes break how the container operates.

* Version bump

* Update README.md