feat: Use command probe for increased accuracy (#344) (#359)
The liveness & readiness probes previously checked that the web server
was up, but did not check that FTL was running. There are conditions
under which the web server starts, but FTL fails to start. Using HTTP
probes does not catch those conditions, but the web server offers an
endpoint that returns the DNS status. If the web server is not up, this
probe will continue to fail since it can't `curl` the endpoint. In
addition, it will now also fail if the DNS server is down.

Signed-off-by: javex <code@inexplicity.de>